uploader/requirements.txt
pyserial~=3.4
tqdm~=4.46.1
